added sam integration + lambda function cleanup
