+++ lambda/deploy.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+sam build && sam deploy --guided
