+++ lambda/samconfig.toml
@@ -1,0 +1,9 @@
+version = 0.1
+[default]
+[default.deploy]
+[default.deploy.parameters]
+stack_name = "twitter-tweets-analysis"
+s3_bucket = "aws-sam-cli-managed-default-samclisourcebucket-16y62ke5wvr8x"
+s3_prefix = "twitter-tweets-analysis"
+region = "us-east-1"
+capabilities = "CAPABILITY_IAM"
