+++ lambda/template.yaml
@@ -1,0 +1,44 @@
+AWSTemplateFormatVersion: '2010-09-09'
+Transform: AWS::Serverless-2016-10-31
+Description: Twitter Tweets Analysis (Serverless Resources)
+
+# More info about Globals: https://github.com/awslabs/serverless-application-model/blob/master/docs/globals.rst
+Globals:
+    Function:
+        Timeout: 3
+        Environment:
+            Variables:
+                KEY: hpAYD8Shnb2B2OHnCEUxQRSHu
+                SECRET: Bg8a3o80jHf1gqX0OuK77owBSLpaYYhayhysHYgGOfj6Oma38S
+
+Resources:
+    AnalyzeTweetsFunction:
+        Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
+        Properties:
+            CodeUri: analyzeTweets/
+            Handler: app.lambda_handler
+            Runtime: python3.8
+            Policies:
+                # Give permission for Comprehend + Lambda
+                - ComprehendFullAccess
+                - AWSLambdaExecute
+            Events: # More info about API Event Source: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#api
+                AnalyzeTweets: 
+                    Type: Api # More info about API Event Source: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#api
+                    Properties:
+                        Path: /analyze/tweets
+                        Method: get
+
+Outputs:
+    # ServerlessRestApi is an implicit API created out of Events key under Serverless::Function
+    # Find out more about other implicit resources you can reference within SAM
+    # https://github.com/awslabs/serverless-application-model/blob/master/docs/internals/generated_resources.rst#api
+    AnalyzeTweetsApi:
+        Description: "API Gateway endpoint URL for Prod stage for Analyze Tweets function"
+        Value: !Sub "https://${ServerlessRestApi}.execute-api.${AWS::Region}.amazonaws.com/Prod/analyze/tweets"
+    AnalyzeTweetsFunction:
+        Description: "Analyze Tweets Lambda Function ARN"
+        Value: !GetAtt AnalyzeTweetsFunction.Arn
+    AnalyzeTweetsFunctionIamRole:
+        Description: "Implicit IAM Role created for Analyze Tweets function"
+        Value: !GetAtt AnalyzeTweetsFunction.Arn
